intelligence/automation/requirements.txt
functions-framework>=3.5.0
google-cloud-bigquery>=3.13.0
simple-salesforce>=1.12.6
hubspot-api-client>=7.0.0

